Fixed an off-by-one error in the coordinate mapping code.
I have good reason to believe this is a bug that has been present for many
months now . . . for some reason, I was treating a point at 100% to the right
as coordinate s, rather than s-1, where s is the width of the image.

This should help a little with the ongoing process of debugging.