The same issue is still occuring . . .
For some reason, the visualizations occasionally freeze. They stop updating --
but, in reality, visualizations are still being merged, still being rendered
. . . which seems slightly odd.

Something odd is happening; I would guess that it is a race condition of some
sort. I cannot seem to pin down exactly where and why the freeze is occuring;
I suppose that having a little bit of context would be a good idea . . .