I am seriously reconsidering the possiblity of introducing the Canvas again.
Having an additional layer of abstraction between the rendering and the
painting would probably help quite a bit; the only question is if the
additional memory usage would actually be worth the cost.

Either that, or I need a better system of calculating the maximum range a
visualization's viewport can view.

Out of the two, I think the latter is a better idea . . .