Bug discovered with the current informer setup.
I apparently forgot that the order modules are destructed has nothing to do
with their dependencies upon each other; the informer can be destructed before
all the other modules, for example. When this is the case, then data cannot be
send from module destructors, which can be rather annoying at times.