Something to think about is module IDs. How are modules distinguished?
The original system of assigning module IDs in a round-robin form does work,
but it seems to be slightly . . . hmm. Hackish? It references modules by name,
which is clearly not optimal for i18n purposes. Referencing modules by some
attribute would be a much better idea. Since all module configurations are
loaded upon commencement of execution, it is quite clear how this could
operate.

If every module is assigned a UUID inside its configuration file (perhaps a
separate configuration file called "uuid.conf" or something) that is generated
as a hash of the version and platform of the module, then logarithmic lookups
are available.

However, this is not good enough. If one has 500,000 packets per second, one
does not exactly wish to perform 500,000 logarithmic lookups per second, even
on a data set as small as thirty elements. It is simply . . . not done. Using
a hash table could work in this situation, though.

The requirements are then, I suppose:
- unique to each module version
- unique to each platform
- constant-time lookups by ID
- same across multiple compilations (e.g. not random and not UU)

It is clear, from those requirements, that a hashing function is required.