I've been doing quite a bit of thinking recently.
Enough that I've condensed my thoughts down into several pages of notes. These
notes are, sadly enough, not in an electronic format. As a result, much of my
thinking is not available for redistribution.

I will, from time to time, copy the salient points into the git log, however.

Now is unfortunately not such a time. I'd like to code instead.