Fixed a few bugs in ZoneReader.
The "polisher", as it is currently called, may need something of a redesign.
There is an additional duty that I would like to assign to it, or rather a
feature that I would like modules to have that fits best in the monitor
module.

The ability to automatically generate some configuration values from the
environment would be rather nice. Rather than having to hardcode or
precalculate calloc()'s offset, instead it could be calculated at run-time
before the target process is launched.

Regardless, it still needs to be rewritten.