Slightly off-topic, but eh.
A system that I can use to find the address that libc is loaded into with
almost perfect accuracy is basically to look through /proc/self/maps to find
anything of the form '/libc' followed by a non-alphabetic character. If no
libraries of this sort are found (or multiple), then die and ask the user to
specify the actual path on the command-line. This can be done in the monitor,
and the final path simply passed onto the informer via a configuration item
anyhow.

This should be portable and reliable enough. The only other requirement is the
offset of the function to resolve, such as calloc(). (0x75e50 in glibc 2.11,
by the way.) From there, the function pointer is trivial.