Began implementing an Analyzer framework.
In essence, the Analyzer framework should allow modules to actually gather
information about the target and its environment without doing most of the
processing themselves.

The two main Analyzers that I have in mind are an ELF analyzer and a DWARF
analyzer. The DWARF analyzer I will leave for the moment, though . . .