Implemented support for waiting for a specific number of client connections.
One rather, shall we say . . . pressing issue: the naming scheme of global
functions in the common/ include directory.

Functions defined in the Informer header file have AI_ prepended to the
function name, which seems to work out quite well. Module-specific functions
have the AM_ prefix; as do the Marshaller functions.

But functions that will be used both in the Informer and also in the monitor
(and possibly also in the visualizer) -- what prefix should they have?