One possible system for rendering . . .
It would be possible to adapt a QGraphicsView to render the visualizations if
necessary. This would probably be faster (and have less bugs) than any system
that I could personally write.

The only problem is drawing text; by default the text is scaled along with the
other content. I would have to draw the text separately myself.