I am beginning to question the need for a marshaller.
The only real use that I can see for the marshaller is for symbol resolution.
And, if all goes as planned, symbol resolution will be provided by a Monitor-
supplied module, e.g. doing the processing within the Monitor itself. The
symbol data can simply be supplied to the Visualizer for the Visualizer to
use, then an R-tree can be used to do the symbol lookups . . .

If I am correct, and the marshaller is not required, then it will make the
implementation of the monitor (and the modules!) that much cleaner.