Implemented a basic mempool and started using it in the monitor.
The mempool is basically a huge pool of memory of which bits can be reserved,
sort of like a static memory allocator. It is byte-perfect (beware buffer
overruns . . .) but has no support for actually removing elements, thus it is
rather unsuited to many applications.

It does, however, work well for some types of RTree.