Time for a little redesign of the Visualizer.
Basically, the idea is that the current system of using one DataStore per
module across every DataInput is kind of . . . not a good one.

The new plan is to use one DataStore per DataInput and set up a composite
DataStore that Viewports can access. That way, it is possible to set up a
visualization that only takes data from a selection of data sources.

Surprisingly, only minor modifications to the framework will be required. The
module artisans, of course, will need to be changed dramatically, but still.

Better now than later.