Next challenge: implement a good system such that the monitor terminates.
Essentially, everything seems to be in place for a nice setup, except for one
thing. The monitor currently has to be killed with a signal before it
terminates. Clearly not optimal.

The problem is that I never really designed the whole zone system around such
a criterion. I need to think about what would be best here.