Something that bears thought: an interface for retrieval from DataStores.
It seems to me that implementing multi-threadedness would be a very good idea.
To that end, I've been thinking about implementing a sort of "DataStoreRequest"
system. One creates an instance of a particular subclassed DataStoreRequest,
and then gives the CompositeDataStore the request. The request is then placed
on a queue . . . which is really ugly and smacks of a little bit too much
asynchronity for my taste.

Perhaps duplicating the QReadWriteLocker using POSIX semaphores would be a good
idea? Then multi-threading support is built-in . . .

Hmm, actually, that sounds like a very good idea.