Implemented Mempool implementation. Kind of hackish and wasteful but it works.
The mempool essentially works by rounding the size of the element to allocate
up to the nearest power of two, then allocating a chunk of memory of that size
from a large buffer. The mempool can be configured at compile-time via the
configuration items poolSize and poolSizeThreshold in the build configuration
file.